Add missing db DSN flag and refactor code
Add missing flag for db DSN so that database URL can be configured. Move
the db config instantiation out to config. This reduces irrelevant code
in the recorder constructor and makes the constructor easier to read.

Do not start the DB connection pool until the server is started and
clean up the pool after shutdown.

Add missing validation tests that make use of the test data.

Improve logging by reducing duplicate properties and consistent
terminology in messages. Reduce log level for messages that are likely
to be verbose.