Fix bugs in event validation and make criteria stronger
Fix a bugs where:
 * empty events in batch were considered valid
 * cid was not checked for validity
 * peer ID was not checked for validity
 * time span of start and event time was unchecked.

Refactor codes to instead check the zero valued struct for validity, as
we need to do that anyway. Using pointers doesn't make sense here
because all the pointer fields are required to have valid content and we
do not need to infer whether they were actually present in JSON or not.

Add extra tests to cover more edge cases